% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{add2Path}
\alias{add2Path}
\title{Adds a variable to the global search path of the current environment}
\usage{
add2Path(newPath)
}
\arguments{
\item{newPath}{the path that is added}
}
\description{
Adds a variable to the global search path of the current environment
}
\examples{
\dontrun{
# add path
addPath("pathtosomewhere")
}
}
\author{
Chris Reudenbach
}

