% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{checkPC}
\alias{checkPC}
\title{Adds a variable to the global search path of the current environment}
\usage{
checkPC(binaryName = NULL, checkName = "PCRZP")
}
\arguments{
\item{binaryName}{name of binary sofware "saga", "otb"}

\item{checkName}{name of PC}
}
\description{
Adds a variable to the global search path of the current environment
}
\examples{
\dontrun{
# add path
checkPC("saga",checkName="PCRZP")
}
}
\author{
Chris Reudenbach
}

