% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2giKarim.R
\name{checkPCDomain}
\alias{checkPCDomain}
\title{Checks if running on a specified computer domain}
\usage{
checkPCDomain(cliCode = NULL, prefixPC = "PCRZP")
}
\arguments{
\item{cliCode}{code of the sofware currently \code{saga} and \code{otb} are supported}

\item{prefixPC}{contains the an arbitrary part of the computer name. It always starts with the first letter.}
}
\description{
Checks if the computer belongs to the Marburg Universitys computer pools
}
\examples{
\dontrun{
# add path
checkPCDomain("saga",prefixPC="PCRZP")
}
}
\author{
CR
}
\keyword{internal}
