% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{createFolders}
\alias{createFolders}
\alias{createFolder}
\title{Compile folder list and create folders}
\usage{
createFolders(root_folder, folders, create_folders = TRUE)
}
\arguments{
\item{root_folder}{root directory of the project.}

\item{folders}{list of subfolders within the project directory.}

\item{create_folders}{create folders if not existing already.}
}
\value{
List with folder paths and names.
}
\description{
Compile folder list with absolut paths and create folders if
necessary.
}
\examples{
\dontrun{
 createFolders(root_folder = tempdir(), folders = c('data/', 'data/tmp/'))
}
# Create folder list and set variable names pointing to the path values
}
