% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{createMeta}
\alias{createMeta}
\title{Create list of metadata from project environment.}
\usage{
createMeta(prj_name)
}
\arguments{
\item{prj_name}{name of the project}
}
\value{
list of metadata.
}
\description{
Create list of metadata from project environment.
}
\examples{
\dontrun{
createMeta(tempdir())
}

}
\keyword{internal}
