% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{enviLoad}
\alias{enviLoad}
\title{Load data from rds format and associated yaml metadata file.}
\usage{
enviLoad(file_path)
}
\arguments{
\item{file_path}{name and path of the rds file.}
}
\value{
list of 2 containing data and metadata.
}
\description{
Load data from rds format and associated yaml metadata file.
}
\examples{
\dontrun{
a <- 1
meta <- list(a = "a is a variable")
enviSave(a, file.path(tempdir(), "test.rds"), meta)
b <- enviLoad(file.path(tempdir(), "test.rds"))
}

}
