% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otb_linux_search.R
\name{findOTB}
\alias{findOTB}
\title{Locate Orfeo ToolBox (OTB) installations}
\usage{
findOTB(searchLocation = NULL, quiet = TRUE)
}
\arguments{
\item{searchLocation}{Character. On Linux: mountpoints/roots to search.
If NULL, defaults to "default" which expands to c("~","/opt","/usr/local","/usr").}

\item{quiet}{Logical.}
}
\value{
data.frame of installations or FALSE.
}
\description{
Dispatcher that calls the OS-specific search function and returns a
normalized installations table.
}
