% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalRun.R
\name{gdal_help}
\alias{gdal_help}
\title{Get and cache "--help" output for a GDAL command}
\usage{
gdal_help(gdal, cmd)
}
\arguments{
\item{gdal}{A gdal_context.}

\item{cmd}{Utility name, e.g. "gdalwarp".}
}
\value{
Character scalar (full help text).
}
\description{
Get and cache "--help" output for a GDAL command
}
