% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{getGrassParams4W}
\alias{getGrassParams4W}
\title{Initialize the enviroment variables on a Windows OS}
\usage{
getGrassParams4W(setDefaultGrass = NULL, DL = "C:", verSelect = FALSE)
}
\arguments{
\item{setDefaultGrass}{default = NULL wil force a search for 'GRASS GIS' You may
provide a valid combination as c("C:\\OSGeo4W64","grass-7.0.5","osgeo4w")}

\item{DL}{raster or sp object}

\item{verSelect}{if TRUE you must interactivley selcect between alternative installations}
}
\value{
getGrassParams4W initializes the usage of GRASS7.
}
\description{
Initialize the enviroment variables on a Windows OS for using 
 'GRASS GIS' via \link{rgrass7}
}
\details{
The concept is very straightforward but for an all days usage pretty
 helpful. You need to provide a GDAL conform raster file, a \link{raster}
 object. This settings will be used to initialize a temporary but static
 \href{https://CRAN.R-project.org/package=rgrass7}{rgrass7} environment. During the rsession you will have full access to
 GRASS7 via the wrapper packages. .
}
\examples{
\dontrun{
# automatic retrieval of valid 'GRASS GIS' environment settings 
# if more than one is found the user has to choose.
getGrassParams4W()

# typical standalone installation
getGrassParams4W(c("C:\\\\Program Files\\\\GRASS GIS 7.0.5","GRASS GIS 7.0.5","NSIS"))

# typical OSGeo4W64 installation
getGrassParams4W(c("C:\\\\OSGeo4W64","grass-7.0.5","osgeo4W"))
}
}

