% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{getSimpleClass}
\alias{getSimpleClass}
\title{Checks if x is a raster or sp object}
\usage{
getSimpleClass(obj)
}
\arguments{
\item{obj}{R raster* or sp object}
}
\description{
Checks if x is a raster or sp object
}
\examples{
\dontrun{
# add path
getSimpleClass(x)
}
}
\author{
Chris Reudenbach
}

