% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{getSpatialClass}
\alias{getSpatialClass}
\title{Checks if x is a raster or sp object}
\usage{
getSpatialClass(obj)
}
\arguments{
\item{obj}{R raster* or sp object}
}
\description{
Checks if x is a raster or sp object
}
\examples{
\dontrun{
# add path
getSpatialClass(x)
}
}
\author{
Chris Reudenbach
}

