% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initGRASS.R
\name{initGRASS}
\alias{initGRASS}
\title{The function initGRASS setup enviroment for using \link{rgrass7} with GRASS}
\usage{
initGRASS(x = NULL, setDefaultGrass = NULL, SP = NULL)
}
\arguments{
\item{x}{raster or sp object}

\item{setDefaultGrass}{default = NULL will force a search for GRASS You
may provide a valid combination as c("C:\\OSGeo4W64","grass-7.0.5","osgeo4w")}

\item{SP}{searchpath}
}
\value{
initGRASS initializes the usage of GRASS7.
}
\description{
Function that initializes environment and pathes for GRASS7x. '
Despite the GRASS GIS seup is performed by the initGRASS() funtion of the
\link{rgrass7} package, there are some workarounds necessary. 
While initGRASS works fine for known pathes and environmental varibles, one 
will find that the integration of Windows based GRASS especially as provided 
by OSGeo4W or the usage of parallel installations could be cumbersome. 
initGRASS trys to find valid GRASS binaries by analyzing the initial GRASS script files.
If necessary it set the system variables and finally it initialize GRASS for R with user
provided  valid raster or sp object.\cr\cr 
*NOTE* If you have more than one valid installation you will be ask to select.
}
\details{
The concept is very straightforward but for an all days usage pretty helpful. 
You need to provide a \link{raster}/\link{sp} spatial object which is correct georeferenced
The resulting params will be used to initialize a temporary but static 
\href{https://CRAN.R-project.org/package=rgrass7}{rgrass7}
environment. During the rsession you will have full access to
GRASS7 via the \link{rgrass7} wrapper package.
}
\examples{
\dontrun{
# get meuse data
library(sp)
data(meuse) 
coordinates(meuse) <- ~x+y 
proj4string(meuse) <-CRS("+init=epsg:28992") 

# automatic search and find of GRASS binaries if 
# more than one you have to choose. 
initGRASS(meuse) 

# assuming a typical standalone installation 
initGRASS(meuse,c("C:\\\\Program Files\\\\GRASS GIS7.0.5","GRASS GIS 7.0.5","NSIS")) 

# assuming a typical OSGeo4W installation
initGRASS(meuse,c("C:\\\\OSGeo4W64","grass-7.0.5","osgeo4W"))

# string for Linux c("/usr/bin","grass72") '
}
}
\author{
Chris Reudenbach
}

