% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOTB.R
\name{initOTB}
\alias{initOTB}
\title{The function initOTB set up OTB bindings}
\usage{
initOTB(binPathOtb = NULL, rootPathOtb = NULL, otbType = NULL,
  DL = "C:", selectOtbVer = FALSE)
}
\arguments{
\item{binPathOtb}{string contains path to where the otb binaries are located}

\item{rootPathOtb}{string provides the root folder of the \code{binPathOtb}}

\item{otbType}{string}

\item{DL}{string hard drive letter default is \code{C:}}

\item{selectOtbVer}{boolean default is FALSE. If there is more than one OTB installation and \code{selectOtbVer} = TRUE the user can select interactively the preferred OTB version}
}
\value{
add otb pathes to the enviroment and creates global variables otbPath
}
\description{
The function initOTB trys to locate all valid OTB installation
 and returns the pathes and environment settings. All valid means that it
 looks for the \code{otb_cli.bat} file. If the file is found in a \code{bin} folder it is assumed to be a valid OTB binary installation.
}
\details{
if called without any parameter \code{initOTB()} it performs a full search over the hardrive \code{C:}. If it finds one or more OTB binaries it will take the first hit. You have to set \code{selectOtbVer = TRUE} for an interactive selection of the preferred version.
}
\note{
It is strongly recommended to set the path manually. Using a osgeo4w installation it is typically \code{C:/OSGeo4W64/bin/}
}
\examples{
\dontrun{
# call it for a default OSGeo4W installation of the OTB
initOTB("C:/OSGeo4W64/bin/")

# call it for a default Linux installation of the OTB
initOTB("/usr/bin/")
}
}
\author{
CR
}

