% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSAGA.R
\name{initSAGA}
\alias{initSAGA}
\title{initSAGA setup SAGA binaries}
\usage{
initSAGA(defaultSAGA = NULL, DL = "C:", MP = "/usr")
}
\arguments{
\item{defaultSAGA}{string contains path to SAGA binaries}

\item{DL}{drive letter}

\item{MP}{mount point}
}
\value{
add saga pathes to the enviroment and creates global variables sagaPath, sagaModPath and sagaCmd
}
\description{
initSAGA setup SAGA binaries
}
\examples{
\dontrun{

# full search
initSAGA()

typical OSGeo4W64 installation 
initSAGA(c("C:\\\\OSGeo4W64\\\\apps\\\\saga","C:\\\\OSGeo4W64\\\\apps\\\\saga\\\\modules"))
}
}

