% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2GI-package.R
\docType{package}
\name{link2GI}
\alias{link2GI}
\alias{link2GI-package}
\title{Bridges to the GI-World}
\description{
A simple wrapper for linking  GI/RS functionality to R
}
\details{
Functions for linking  GI/RS functionality to R
}
\note{
It is important to keep in mind that all binaries need to be installed correctly on your system.
the link2GI package just tries to generate correct environment settings as system and path variables for the most of the known issues. 
It is tested under 'Windows 7' and 'Windows 10' as well as on the Ubuntu/Debian 'Linux' distributions. 
The 'OSX' operation system should run but is not tested.
}
\author{
Chris Reudenbach Tim Appelhans
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}
