% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{makeGlobalVariable}
\alias{makeGlobalVariable}
\title{Generates a variable with a certain value in the R environment}
\usage{
makeGlobalVariable(names, values)
}
\arguments{
\item{names}{vector with the  names of the variable(s)}

\item{values}{vector with values of the variable(s)}
}
\description{
Generates a variable with a certain value in the R environment.
}
\examples{
\dontrun{
# creates the global variable \\code{path_data} with the value \\code{~/data}
makeGlobalVariable(names = "path_data", values = "~/data")
}

}
\keyword{internal}
