% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOTB.R
\name{mkOTBcmd}
\alias{mkOTBcmd}
\title{paste the OTB command list into a system call compatible string}
\usage{
mkOTBcmd(path_OTB, otb_algorithm, algo_cmd)
}
\arguments{
\item{path_OTB}{path to the OTB installastion directory as provided by linkOTB()}

\item{otb_algorithm}{the currently choosen otb algorithm keyword}

\item{algo_cmd}{the modified algorithm parameter list}
}
\description{
helper function which paste the OTB command list into a system call compatible string
}
