% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbIntro.R
\name{otb_args_spec}
\alias{otb_args_spec}
\title{Return normalized OTB parameter spec for an application}
\usage{
otb_args_spec(algo, gili = NULL)
}
\arguments{
\item{algo}{Character. OTB application name.}

\item{gili}{Optional list returned by [linkOTB()]. If `NULL`, [linkOTB()] is called.}
}
\value{
data.frame with columns: key, type, mandatory, has_pixel, pixel_default, has_default, default, class, desc
}
\description{
Return normalized OTB parameter spec for an application
}
