% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbIntro.R
\name{otb_optional}
\alias{otb_optional}
\title{Get optional parameters for an OTB application (keys + defaults optional)}
\usage{
otb_optional(algo, gili = NULL, with_defaults = TRUE)
}
\arguments{
\item{algo}{Character. OTB application name.}

\item{gili}{Optional list returned by [linkOTB()]. If `NULL`, [linkOTB()] is called.}

\item{with_defaults}{Logical. If TRUE, fill parsed defaults where available; else NA.}
}
\value{
Named list.
}
\description{
Get optional parameters for an OTB application (keys + defaults optional)
}
