% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbIntro.R
\name{otb_show}
\alias{otb_show}
\title{Print a quick console view of an OTB application's parameters}
\usage{
otb_show(algo, gili = NULL)
}
\arguments{
\item{algo}{Character. OTB application name.}

\item{gili}{Optional list returned by [linkOTB()]. If `NULL`, [linkOTB()] is called.}
}
\value{
Invisibly list(caps=..., spec=...)
}
\description{
Print a quick console view of an OTB application's parameters
}
