% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGvec.R
\name{readGvec}
\alias{readGvec}
\title{Read GRASS 7 vector into  sf object}
\usage{
readGvec(x, objName, gisdbase, location)
}
\arguments{
\item{x}{sf* object corresponding to the settings of the corresponding GRASS container}

\item{objName}{name of GRASS layer}

\item{gisdbase}{GRASS gisDbase folder}

\item{location}{GRASS location name containing \code{objName)}}
}
\description{
Read GRASS 7 vector into  sf object
}
\examples{
\dontrun{
## example 
# get meuse data as sf object
require(sf)
meuse_sf = st_as_sf(meuse, 
                   coords = c("x", "y"), 
                   crs = 28992, 
                   agr = "constant")
    

readGvec(x = meuse_sf,
          objName = "meuse_R-G",
          gisdbase = "~/temp3",
          location = "project1")
}
}
\author{
Chris Reudenbach
}

