% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2GI-package.R
\docType{data}
\name{rgb}
\alias{rgb}
\title{RGB ortho-image from an arbitray Marburg Open Forest (MOF) plot}
\format{\code{"raster::raster"}}
\description{
Example data set containing a RGB ortho-image of a small plot sampled in the Maburg University Forest aka Marburg Open Forest (MOF). The resolution is 10 cm, projection ETRS89 UTM32
}
