% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{saveEnvi}
\alias{saveEnvi}
\title{Saves data in rds format and adds a yaml metadata file.}
\usage{
saveEnvi(variable, file_path, meta)
}
\arguments{
\item{variable}{name of the data variable to be saved.}

\item{file_path}{name and path of the rds file.}

\item{meta}{name of the metadata list.}
}
\description{
Saves data in rds format and saves metadata in a corresponding yaml file.
}
\examples{
\dontrun{
a <- 1
meta <- list(a = "a is a variable")
saveEnvi(a, file.path(tempdir(), "test.rds"), meta)
}

}
