% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalControls.R
\name{searchGDALW}
\alias{searchGDALW}
\title{Search recursively for valid 'GDAL' installation(s) on a 'Windows' OS}
\usage{
searchGDALW(DL = "C:", quiet = TRUE)
}
\arguments{
\item{DL}{drive letter default is "C:"}

\item{quiet}{boolean  switch for supressing console messages default is TRUE}
}
\value{
A dataframe with the 'GDAL' root folder(s) the version name(s) and the installation type(s).
}
\description{
Search for valid 'GDAL' installations on a 'Windows' OS
}
\examples{

run = FALSE
if (run) {
# get all valid GDAL installation folders and params
searchGDALW()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
