% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchOSgeo4WGrass}
\alias{searchOSgeo4WGrass}
\title{search for valid OSGeo4W GRASS installation(s) on a given Windows drive}
\usage{
searchOSgeo4WGrass(DL = "C:")
}
\arguments{
\item{DL}{drive letter to be searched, default is "C:"}
}
\value{
dataframe with the GRASS root dir, version name and installation type code word
}
\description{
Provides an  estimation of valid GRASS installation(s) on your Windows system. There is a major difference between osgeo4W and standalone installations. The functions trys to find all valid installations by analysing the calling batch scripts.
}
\details{
searchOSgeo4WGrass
}
\examples{
\dontrun{
# get all valid GRASS installation folders and params on WindowsOS
grassParam<- searchOSgeo4WGrass()
}
}
\author{
Chris Reudenbach
}

