% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otb_linux_search.R
\name{searchOTBX}
\alias{searchOTBX}
\title{Search recursively for valid OTB installation(s) on Linux/macOS}
\usage{
searchOTBX(MP = "default", quiet = TRUE)
}
\arguments{
\item{MP}{Character. "default" expands to c("~","/opt","/usr/local","/usr").
Otherwise, one or more mountpoints/roots.}

\item{quiet}{Logical.}
}
\value{
data.frame with OTB installations (or FALSE).
}
\description{
Searches for an executable `otbcli` and validates presence of
`otbApplicationLauncherCommandLine` in the same bin folder (or base/bin).
}
\keyword{internal}
