% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchSAGA4W}
\alias{searchSAGA4W}
\title{Search for valid 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGA4W(DL = "C:")
}
\arguments{
\item{DL}{drive letter default is "C:"}
}
\value{
a dataframe with the 'SAGA GIS' root folder, the version name and the installation type
}
\description{
Search for valid 'SAGA GIS' installation(s) on a given 'Windows' drive
}
\examples{
\dontrun{
#### Examples how to use searchSAGA4W 

# get all valid SAGA installation folders and params
sagaParams<- searchSAGA4W()
}
}
\author{
Chris Reudenbach
}

