% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalControls.R
\name{setenvGDAL}
\alias{setenvGDAL}
\title{Usually for internally usage, initializes and set up access to the 'GDAL' command line interface}
\usage{
setenvGDAL(bin_GDAL = NULL)
}
\arguments{
\item{bin_GDAL}{Character. Path to GDAL bin directory OR full path to gdalinfo(.exe).}
}
\value{
The normalized GDAL bin directory path (invisibly).
}
\description{
Initializes and sets up access to the 'GDAL' command line interface
}
\examples{
\dontrun{
setenvGDAL("C:/OSGeo4W64/bin")
}
}
\keyword{internal}
