% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbControls.R
\name{setenvOTB}
\alias{setenvOTB}
\title{Usually for internally usage, initializes and set up  access to the 'OTB' command line interface}
\usage{
setenvOTB(bin_OTB = NULL, root_OTB = NULL)
}
\arguments{
\item{bin_OTB}{string contains the path to the 'OTB' binaries}

\item{root_OTB}{string contains the full string to the root folder
containing the 'OTB' installation'}
}
\value{
Adds 'OTB' pathes to the environment and creates the variable global string variable \code{otbCmd}, that contains the path to the 'OTB' binaries.
}
\description{
Initializes and set up  access to the 'OTB' command line interface
}
\examples{
\dontrun{
## example for the most common default OSGeo4W64 installation of OTB
setenvOTB(bin_OTB = 'C:\\\\OSGeo4W64\\\\bin\\\\',
          root_OTB = 'C:\\\\OSGeo4W64')
}
}
\keyword{internal}
