% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGVec.R
\name{writeGvec}
\alias{writeGvec}
\title{Write sf object to GRASS 7}
\usage{
writeGvec(x, objName, gisdbase, location)
}
\arguments{
\item{x}{sp* or sf* object}

\item{nrasters}{vector of channels to use from x. Default =nlayers(x)}
}
\description{
Write sf object to GRASS 7
}
\examples{
\dontrun{
## example 
# get meuse data as sf object
require(sf)
meuse_sf = st_as_sf(meuse, 
                   coords = c("x", "y"), 
                   crs = 28992, 
                   agr = "constant")
    

writeGvec(x = meuse_sf,
          objName = "meuse_R-G",
          gisdbase = "~/temp3",
          location = "project1")
}
}
\author{
Chris Reudenbach
}

