% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{checkPCRZP}
\alias{checkPCRZP}
\title{Checks if the computer belongs to the Marburg Universitys computer domain}
\usage{
checkPCRZP(cliCode = NULL, prefixPC = "PCRZP")
}
\arguments{
\item{cliCode}{code of the sofware currently "saga" and "otb" are supported}

\item{prefixPC}{name of PC}
}
\description{
Checks if the computer belongs to the Marburg Universitys computer pools
}
\examples{
\dontrun{
# add path
checkPCRZP("saga",prefixPC="PCRZP")
}
}
\author{
CR
}

