% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{createFolders}
\alias{createFolders}
\alias{createFolder}
\title{Compile folder list and create folders}
\usage{
createFolders(
  root_folder,
  folders,
  folder_names = NULL,
  path_prefix = NULL,
  create_folders = FALSE
)
}
\arguments{
\item{root_folder}{root directory of the project.}

\item{folders}{list of subfolders within the project directory.}

\item{folder_names}{names of the variables that point to subfolders. If not
provided, the base paths of the folders is used.}

\item{path_prefix}{a prefix for the folder names.}

\item{create_folders}{create folders if not existing already.}
}
\value{
List with folder paths and names.
}
\description{
Compile folder list with absolut paths and create folders if
necessary.
}
\examples{
\dontrun{
# createFolders(root_folder = "~/edu", folders = c("data/", "data/tmp/"))
}
# Create folder list and set variable names pointing to the path values
}
