% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalControls.R
\name{findGDAL}
\alias{findGDAL}
\title{Search recursivly existing 'GDAL binaries' installation(s) at a given drive/mountpoint}
\usage{
findGDAL(searchLocation = "default", quiet = TRUE)
}
\arguments{
\item{searchLocation}{drive letter to be searched, for Windows systems default
is \code{C:}, for Linux systems default is \code{/usr}.}

\item{quiet}{boolean  switch for supressing console messages default is TRUE}
}
\value{
A dataframe with the 'GDAL' root folder(s),  and command line executable(s)
}
\description{
Provides an  list of valid 'GDAL' installation(s) 
on your 'Windows' system. There is a major difference between osgeo4W and 
stand_alone installations. The functions trys to find all valid 
installations by analysing the calling batch scripts.
}
\examples{

run = FALSE
if (run) {
# find recursively all existing 'GDAL' installations folders starting 
# at the default search location
findGDAL()
}
}
\author{
Chris Reudenbach
}
