% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassControl.R
\name{findGRASS}
\alias{findGRASS}
\title{Returns attributes of valid 'GRASS GIS' installation(s) on the system.}
\usage{
findGRASS(searchLocation = "default", ver_select = FALSE, quiet = TRUE)
}
\arguments{
\item{searchLocation}{On Windows MUST start with drive letter + colon, e.g. "C:", "C:/", "C:/Users/...".
Defaults to "C:/". On Unix defaults to "/usr/bin".}

\item{ver_select}{If TRUE and more than one installation is found, interactively select one.}

\item{quiet}{Suppress messages.}
}
\value{
FALSE or data.frame(instDir, version, installation_type)
}
\description{
Retrieve a list of valid 'GRASS GIS' installation(s) on your system.
On Windows, uses searchGRASSW() (cmd-free). On Unix, uses searchGRASSX().
}
