% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{findGRASS}
\alias{findGRASS}
\title{Search recursivly valid 'GRASS GIS' installation(s) at a given drive/mountpoint}
\usage{
findGRASS(searchLocation = "default")
}
\arguments{
\item{searchLocation}{drive letter to be searched, for Windows systems default
is \code{C:}, for Linux systems default is \code{/usr}.}
}
\value{
A dataframe with the 'GRASS GIS' root folder(s), version name(s) and 
installation type code(s)
}
\description{
Provides an  list of valid 'GRASS GIS' installation(s) 
on your 'Windows' system. There is a major difference between osgeo4W and 
stand_alone installations. The functions trys to find all valid 
installations by analysing the calling batch scripts.
}
\examples{
\dontrun{
# find recursively all existing 'GRASS GIS' installation folders starting at the default search location
findGRASS()
}
}
\author{
Chris Reudenbach
}
