% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbControls.R
\name{findOTB}
\alias{findOTB}
\title{Search recursivly existing 'Orfeo Toolbox' installation(s) at a given drive/mountpoint}
\usage{
findOTB(searchLocation = "default", quiet = TRUE)
}
\arguments{
\item{searchLocation}{drive letter to be searched, for Windows systems default
is \code{C:}, for Linux systems default is \code{/usr}.}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\value{
A dataframe with the 'OTB' root folder(s),  and command line executable(s)
}
\description{
Provides an  list of valid 'OTB' installation(s) 
on your 'Windows' system. There is a major difference between osgeo4W and 
stand_alone installations. The functions trys to find all valid 
installations by analysing the calling batch scripts.
}
\examples{
\dontrun{
# find recursively all existing 'Orfeo Toolbox' installations folders starting 
# at the default search location
findOTB()
}
}
\author{
Chris Reudenbach
}
