% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initgdalUtil.R
\name{initgdalUtils}
\alias{initgdalUtils}
\title{Check and export the gdalUtils settings}
\usage{
initgdalUtils()
}
\value{
a list of the complete capabilities of the current installed gdal version
}
\description{
Check and export the gdalUtils settings. 
You need to have installed the GDAL binaries.
}
\examples{
\dontrun{

# get all available driver 
gdal<- initgdalUtils()

gdal[[1]]$drivers$format_code

# GET BINARY PATH 
gdal[[1]]$path

# get additional and available python tools
gdal[[1]]$python_utilities
}
}
\author{
CR
}

