% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initProj.R
\name{initProj}
\alias{initProj}
\title{Defines and creates folders and variables}
\usage{
initProj(projRootDir = tempdir(), GRASSlocation = "tmp/",
  projFolders = c("data/", "result/", "run/", "log/"))
}
\arguments{
\item{projRootDir}{project github root directory (your github name)}

\item{GRASSlocation}{folder for GRASS data}

\item{projFolders}{list of subfolders in project}
}
\description{
Defines and creates (if necessary) all folders variables
set the SAGA path variables and other system variables
exports all variables to the global environment
}
