% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initSAGA.R
\name{initSAGA}
\alias{initSAGA}
\title{The function initSAGA trys to locate and set up valid SAGA installations}
\usage{
initSAGA(defaultSAGA = NULL, DL = "C:", MP = "/usr")
}
\arguments{
\item{defaultSAGA}{string contains path to SAGA binaries}

\item{DL}{drive letter}

\item{MP}{mount point}
}
\value{
add SAGA pathes to the enviroment and creates global variables sagaPath, sagaModPath and sagaCmd
}
\description{
The function initSAGA trys to locate and set up valid SAGA installations 
 and returns the pathes and correct environment settings. All valid means that it
 looks for the \code{saga_cmd} or \code{saga_cmd.exe} executable of the SAGA CLI. If the file is found it is assumed to be a valid SAGA binary installation.
}
\note{
The SAGA wrapper for R RSAGA is NOT used. The reason is the weird concept of the SAGA developers which are always changing the System calls and so forth. Due to this developing design RSAGA is only running stable with specific versions of SAGA.
}
\examples{
\dontrun{

# full search
initSAGA()

typical OSGeo4W64 installation 
initSAGA(c("C:\\\\OSGeo4W64\\\\apps\\\\saga","C:\\\\OSGeo4W64\\\\apps\\\\saga\\\\modules"))
}
}

