% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2GI-package.R
\docType{package}
\name{link2GI}
\alias{link2GI}
\alias{link2GI-package}
\title{linking R to the GI-world}
\description{
Functions for linking  GI/RS functionality to R
}
\details{
The package provides a simple wrapper for linking  GI/RS functionality to R
}
\author{
Chris Reudenbach
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}

