% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2GI-package.R
\docType{package}
\name{link2GI}
\alias{link2GI}
\alias{link2GI-package}
\title{Linking the GI-World}
\description{
A simple wrapper for linking  GI/RS functionality to R
}
\details{
Functions for linking  GI/RS functionality to R
}
\note{
It is important to keep in mind that all binaries need to be installed correctly on your system. the link2GI package just tries to generate correct settings, system variables and so on for most of the known issues. It is tested for 'Windows 7' and 'Windows 10' as well as for the most popular 'Linux' distributions. The 'OSX' operation system should run but is not tested.
}
\author{
Chris Reudenbach
\cr
\emph{Maintainer:} Chris Reudenbach \email{reudenbach@uni-marburg.de}
}
\keyword{package}

