% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkgdalUtil.R
\name{linkgdalUtils}
\alias{linkgdalUtils}
\title{Search and export the settings of \code{gdalUtils} and 'GDAL' binaries.}
\usage{
linkgdalUtils()
}
\value{
A list of the complete capabilities of the current installed GDAL version
}
\description{
Check and export the \href{https://CRAN.R-project.org/package=gdalUtils}{gdalUtils} and \href{http://www.gdal.org/}{'GDAL'} settings. 
You need to have installed the  'GDAL' binaries.
}
\examples{
\dontrun{

# get all available driver 
gdal<- linkgdalUtils()

gdal[[1]]$drivers$format_code

# GET BINARY PATH 
gdal[[1]]$path

# get additional and available python tools
gdal[[1]]$python_utilities
}
}
\author{
CR
}
