% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkGRASS.R
\name{linkGRASS7}
\alias{linkGRASS7}
\title{Initializes environment variables and pathes for GRASS7}
\usage{
linkGRASS7(x = NULL, defaultGrass = NULL, searchPath = NULL,
  verSelect = FALSE, gisdbase = NULL, mapset = "PERMANENT",
  location = NULL)
}
\arguments{
\item{x}{raster or sp object}

\item{defaultGrass}{if NULL an automatic search will be performed. You may 
also provide a valid combination as c("C:/OSGeo4W64","grass-7.0.5","osgeo4w")}

\item{searchPath}{path or mounting point that will be searched}

\item{verSelect}{boolean if TRUE you may choose interactively the binary version (if found  more than one),  by default FALSE}

\item{gisdbase}{'GRASS' database directory de4fault is \code{tempdir()}}

\item{mapset}{the 'GRASS' mapset}

\item{location}{the 'GRASS' location}
}
\value{
linkGRASS7 initializes the usage of GRASS7.
}
\description{
Initializes the session environment and the system pathes for an easy acces to 
\href{https://grass.osgeo.org/}{'GRASS GIS 7.x'}.  The correct setup of the spatial and projection parameters is
 automatically performed by using either an existing and valid \code{\link{raster}}, \code{\link{sp}} or \code{\link{sf}} object, 
 or manually by providing a list containing the minimum parameters needed.\cr
}
\details{
The concept is very straightforward but for an all days usage pretty 
 helpful. You need to provide a \code{\link{raster}} or \code{\link{sp}} \code{\link{sf}} spatial object
 which has correct spatial and projection properties. These properties are used to initialize either a temporary or a permanent 
 \href{https://CRAN.R-project.org/package=rgrass7}{rgrass7} environment including the correct 'GRASS 7' structure.\cr\cr
 The most time consuming part on 'Windows' Systems is the search process. This can easily take 10 or more minutes. 
 To speed up this process you can also provide a correct parameter set. Best way to do so is to call \code{searchGRASSW} or for 'Linux' \code{searchGRASSX} manually. 
 and call \code{linkGRASS7} with the version arguments of your choice.
}
\note{
'GRASS GIS 7' is excellently supported by the
 \link{rgrass7} wrapper package. Nevertheless 'GRASS GIS' is well known for
 its high demands regarding the correct spatial and reference setup an a bunch 
 of workspace and environment requirements. This becomes even worse on 'Windows' 
 platforms or if several alternative 'GRASS GIS' installations are available.
 If one knows what to do the \code{\link{rgrass7}} package setup function \code{initGRASS} works fine under Linux. 
 This is also valid for well known configurations under the 'Windows' operation system. 
 Nevertheless on university lab or on company computers with restriced privileges and/or using different releases
 like the  \href{http://trac.osgeo.org/osgeo4w/}{'OSGeo4W'} distribution and the  
 \href{https://grass.osgeo.org/download/software/ms-windows/#stand-alone}{stand-alone} installation, 
 or different software releases (e.g. 'GRASS 7.0.5 and GRASS 7.2.0), it becomes often cumbersome or even impossible to get the correct linkages. \cr 
 The function \code{linkGRASS7} tries to find all valid 'GRASS GIS' binaries by analyzing
 the startup script files of 'GRASS GIS'. After identifying the 'GRASS GIS' binaries all
 necessary system variables and settings will be generated and passed to a temporary R enviroment.

If you have more than one valid installation and run \code{linkGRASS7()} without arguments, you will be ask to select one.
}
\examples{
\dontrun{
# get meuse data as sp object
library(link2GI)
require(sp)
data(meuse) 
coordinates(meuse) <- ~x+y 
proj4string(meuse) <-CRS("+init=epsg:28992") 

# get meuse data as sf object
require(sf)
meuse_sf = st_as_sf(meuse, coords = c("x", "y"), crs = 28992, agr = "constant")

# automatic search and find of GRASS binaries if 
# more than one you have to choose. 
linkGRASS7(meuse)
 
# if you do not have any idea if and where GRASS is installed
# Actually this type of linking procedure is highly recommended
linkGRASS7(meuse)

# assuming a typical standalone non-OSGeo4W installation
linkGRASS7(meuse,c("C:/Program Files/GRASS GIS7.0.5","GRASS GIS 7.0.5","NSIS")) 

# assuming a typical OSGeo4W installation
linkGRASS7(meuse,c("C:/OSGeo4W64","grass-7.0.5","osgeo4W"))

# assuming a permanent GRASS folder at "~/temp3" location named "project1" 
# and using an sf object for initialisation
linkGRASS7(meuse_sf,gisdbase = "~/temp3",location = "project1")   

# choose manually the GRASS installation 
linkGRASS7(meuse_sf,verSelect = TRUE)

# choose manually the GRASS installation and change the search location
linkGRASS7(meuse_sf,verSelect = TRUE, searchPath = "D:/")

}
}
\author{
Chris Reudenbach
}

