% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkSAGA.R
\name{linkSAGA}
\alias{linkSAGA}
\title{Locate and bind valid SAGA installation(s)}
\usage{
linkSAGA(defaultSAGA = NULL, DL = "C:", MP = "/usr")
}
\arguments{
\item{defaultSAGA}{string contains path to SAGA binaries}

\item{DL}{drive letter}

\item{MP}{mount point}
}
\description{
Locate and bind valid 'SAGA GIS' \url{http://www.saga-gis.org/}
 installation(s). It returns the pathes and correct environment settings. All
 valid means that it looks for the \code{saga_cmd} or \code{saga_cmd.exe}
 executable of the SAGA CLI. If the file is found it is assumed to be a valid
 SAGA binary installation.
}
\details{
Adds SAGA pathes to the environment and creates the global variables
sagaPath, sagaModPath and sagaCmd
}
\note{
The excellent SAGA wrapper \href{https://CRAN.R-project.org/package=RSAGA}{RSAGA} is NOT used because the
 permanent occucrence of breaking the API calls makes it highly impracticable
 to keep in line. RSAGA fits perfectly your needs if you use 'SAGA GIS'
 (2.0.4 - 2.2.3).
}
\examples{
\dontrun{

# call if you do not have any idea if and where SAGA GIS is installed
linkSAGA()

# typical OSGeo4W64 installation 
linkSAGA(c("C:/OSGeo4W64/apps/saga","C:/OSGeo4W64/apps/saga/modules"))
}
}

