% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link2giKarim.R
\name{makGlobalVar}
\alias{makGlobalVar}
\title{Generates a variable with a certain value in the R environment}
\usage{
makGlobalVar(name, value)
}
\arguments{
\item{name}{character string name of the variable}

\item{value}{character string value of the variable}
}
\description{
Generates a variable with a certain value in the R environment
}
\examples{
\dontrun{

# creates the global var \\code{pathToData} with the value \\code{~/home/data}
makGlobalVar("pathToData","~/home/data") 

}

}
\keyword{internal}
