% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOTB.R
\name{parseOTBAlgorithms}
\alias{parseOTBAlgorithms}
\title{Get OTB modules}
\usage{
parseOTBAlgorithms(gili = NULL)
}
\arguments{
\item{gili}{optional gis linkage as done by `linkOTB()``}
}
\description{
retrieve the OTB module folder content and parses the module names
}
\examples{
\dontrun{
 ## link to the OTB binaries
   otbLinks<-link2GI::linkOTB()
   path_OTB<-otbLinks$pathOTB
 ##p arse all modules
algo<-parseOTBAlgorithms(gili = otbLinks)

##print the list
print(algo)
} 

}
