% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOTB.R
\name{parseOTBAlgorithms}
\alias{parseOTBAlgorithms}
\title{Get OTB modules}
\usage{
parseOTBAlgorithms(gili = NULL)
}
\arguments{
\item{gili}{optional list of avalailable `OTB` binaries if not provided `linkOTB()` is called}
}
\description{
retrieve the OTB module folder content and parses the module names
}
\examples{
\dontrun{
## link to the OTB binaries
otbLink<-link2GI::linkOTB()

 if (otbLink$exist) {

 ## parse all modules
 moduleList<-parseOTBAlgorithms(gili = otbLink)

 ## print the list
 print(moduleList)
 
 } 
}

}
