% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOTB.R
\name{parseOTBFunction}
\alias{parseOTBFunction}
\title{Get OTB function argument list}
\usage{
parseOTBFunction(algo = NULL, gili = NULL)
}
\arguments{
\item{algo}{number or name of the algorithm as provided by `getOTBAlgorithm`}

\item{gili}{optional gis linkage as done by `linkOTB()`}
}
\description{
retrieve the choosen function and parses all arguments with the defaults
}
\examples{
\dontrun{
otbLink<-link2GI::linkOTB()
if (otbLink$exist) {

## parse all modules
algos<-parseOTBAlgorithms(gili = otbLink)


## take edge detection
cmdList<-parseOTBFunction(algo = algos[27],gili = otbLink)
## print the current command
print(cmdList)
}
}
##+##
}
