% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseOTB.R
\name{runOTB}
\alias{runOTB}
\title{Execute the OTB command list via system call}
\usage{
runOTB(otbCmdList = NULL, gili = NULL, retRaster = TRUE, quiet = TRUE)
}
\arguments{
\item{otbCmdList}{the OTB algorithm parameter list}

\item{gili}{optional gis linkage as done by `linkOTB()`}

\item{retRaster}{boolean if TRUE a raster stack is returned}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\description{
Wrapper function which paste the OTB command list into a system call compatible string and execute this command.
}
\details{
Please note that you must check the help if the input argument is set with $inputi or $inputl
}
\examples{
\dontrun{
require(link2GI)
require(raster)
require(listviewer)

## link to OTB
otbLink<-link2GI::linkOTB()

if (otblink$exist) {
 projRootDir<-tempdir()
 data('rgb', package = 'link2GI')  
 raster::plotRGB(rgb)
 r<-raster::writeRaster(rgb, 
                        filename=file.path(projRootDir,"test.tif"),
                        format="GTiff", 
                        overwrite=TRUE)

## for an image output example we use the Statistic Extraction, 
algoKeyword<- "LocalStatisticExtraction"

## extract the command list for the choosen algorithm 
cmd<-parseOTBFunction(algo = algoKeyword, gili = otbLink)

## get help using the convenient listviewer
listviewer::jsonedit(cmd$help)

## define the mandantory arguments all other will be default
cmd$inputi  <- file.path(tempdir(),"test.tif")
cmd$out <- file.path(tempdir(),"test_otb_stat.tif")
cmd$radius <- 7

## run algorithm
retStack<-runOTB(cmd,gili = otbLink)

## plot raster
raster::plot(retStack)

## for a data output example we use the 

algoKeyword<- "ComputeImagesStatistics"

## extract the command list for the choosen algorithm 
cmd<-parseOTBFunction(algo = algoKeyword, gili = otbLink)

## get help using the convenient listviewer
listviewer::jsonedit(cmd$help)

## define the mandantory arguments all other will be default
cmd$inputl  <- file.path(tempdir(),"test.tif")
cmd$ram <- 4096
cmd$out <- file.path(tempdir(),"test_otb_stat.xml")

## run algorithm
retStack<-runOTB(cmd,gili = otbLink)

## plot raster
cat(file.path(tempdir(),"test_otb_stat.xml"))

}
}
}
