% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalControls.R
\name{searchGDALX}
\alias{searchGDALX}
\title{Search recursively for valid 'GDAL' installation(s) on a 'X-based' OS}
\usage{
searchGDALX(MP = "/usr/bin", quiet = TRUE)
}
\arguments{
\item{MP}{drive letter default is "/usr/bin"}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\value{
A dataframe with the 'GDAL' root folder(s) the version name(s) and the installation type(s).
}
\description{
Search for valid 'GDAL' installations on a 'X-based' OS
}
\examples{

run = FALSE
if (run) {
# get all valid GDAL installation folders and params
searchGDALX()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
