% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7Control.R
\name{searchGRASSW}
\alias{searchGRASSW}
\title{Search recursivly valid 'GRASS GIS' installation(s) on a given 'Windows' drive}
\usage{
searchGRASSW(DL = "C:", quiet = TRUE)
}
\arguments{
\item{DL}{drive letter to be searched, default is "C:"}

\item{quiet}{boolean  switch for supressing console messages default is TRUEs}
}
\value{
A dataframe with the 'GRASS GIS' root folder(s), version name(s) and installation type code(s)
}
\description{
Provides an  list of valid 'GRASS GIS' installation(s) on your 'Windows' system. There is a major difference between osgeo4W and stand_alone installations. The functions trys to find all valid installations by analysing the calling batch scripts.
}
\examples{
\dontrun{
# get all valid 'GRASS GIS' installation folders and params at "C:"
searchGRASSW()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
