% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchGRASSX}
\alias{searchGRASSX}
\title{Search for valid 'GRASS GIS' installations at a given Linux mount point}
\usage{
searchGRASSX(MP = "/usr")
}
\arguments{
\item{MP}{default is /usr}
}
\value{
dataframe containing 'GRASS GIS' binary dir(s), version name(s) and installation type code(s)
}
\description{
Search for valid 'GRASS GIS' installations at a given Linux mount point
}
\examples{
\dontrun{
# get all valid 'GRASS GIS' installation folders in the /usr directory (typical location)
grassParam<- searchGRASSX("~/")

# get all valid 'GRASS GIS' installation folders in the home directory
grassParam<- searchGRASSX("~/")
}
}
\author{
Chris Reudenbach
}

