% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass7Control.R
\name{searchGRASSX}
\alias{searchGRASSX}
\title{Search recursivly valid 'GRASS GIS' installation(s) at a given 'Linux' mount point}
\usage{
searchGRASSX(MP = "/usr", quiet = TRUE)
}
\arguments{
\item{MP}{default is /usr}

\item{quiet}{boolean  switch for supressing console messages default is TRUEs}
}
\value{
A dataframe containing 'GRASS GIS' binary folder(s), version name(s) and installation type code(s)
}
\description{
Search for valid 'GRASS GIS' installations at a given 'Linux' mount point
}
\examples{
\dontrun{
# get all valid 'GRASS GIS' installation folders in the /usr directory (typical location)
searchGRASSX("/usr")

# get all valid 'GRASS GIS' installation folders in the home directory
searchGRASSX("~/")
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
