% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchOTBW}
\alias{searchOTBW}
\title{Search for valid 'OTB' installations on a 'Windows' OS}
\usage{
searchOTBW(DL = "C:")
}
\arguments{
\item{DL}{drive letter default is "C:"}
}
\value{
A dataframe with the 'OTB' root dir the Version name and the installation type
}
\description{
Search for valid 'OTB' installations on a 'Windows' OS
}
\examples{
\dontrun{
#### Examples how to use RSAGA and OTB bindings from R

# get all valid OTB installation folders and params
otbParam<- searchOTBW()
}
}
\author{
Chris Reudenbach
}

