% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbControls.R
\name{searchOTBW}
\alias{searchOTBW}
\title{Search recursively for valid 'OTB' installation(s) on a 'Windows' OS}
\usage{
searchOTBW(DL = "C:", quiet = TRUE)
}
\arguments{
\item{DL}{drive letter default is "C:"}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\value{
A dataframe with the 'OTB' root folder(s) the version name(s) and the installation type(s).
}
\description{
Search for valid 'OTB' installations on a 'Windows' OS
}
\examples{
\dontrun{
#### Examples how to use RSAGA and OTB bindings from R

# get all valid OTB installation folders and params
searchOTBW()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
