% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otbControls.R
\name{searchOTBX}
\alias{searchOTBX}
\title{Search recursively for valid 'OTB' installation(s) on a 'Windows' OS}
\usage{
searchOTBX(MP = "/usr", quiet = TRUE)
}
\arguments{
\item{MP}{drive letter default is "C:"}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\value{
A dataframe with the 'OTB' root folder(s) the version name(s) and the installation type(s).
}
\description{
Search for valid 'OTB' installations on a 'Windows' OS
}
\examples{
\dontrun{
# get all valid OTB installation folders and params
searchOTBX()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
