% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchSAGA4W}
\alias{searchSAGA4W}
\title{search for valid SAGA installations on a given Windows drive}
\usage{
searchSAGA4W(DL = "C:")
}
\arguments{
\item{DL}{drive letter default is "C:"}
}
\value{
a dataframe with the SAGA root folder, the version name and the installation type
}
\description{
Provides a fairly good estimation of valid SAGA installation(s) on your Windows system
}
\examples{
\dontrun{
#### Examples how to use searchSAGA4W 

# get all valid SAGA installation folders and params
sagaParams<- searchSAGA4W()
}
}
\author{
Chris Reudenbach
}

