% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchSAGAW}
\alias{searchSAGAW}
\title{Search for valid 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGAW(DL = "C:")
}
\arguments{
\item{DL}{drive letter default is "C:"}
}
\value{
a dataframe with the 'SAGA GIS' root folder, the version name and the installation type
}
\description{
Search for valid 'SAGA GIS' installation(s) on a given 'Windows' drive
}
\examples{
\dontrun{
#### Examples how to use searchSAGAW 

# get all valid SAGA installation folders and params
sagaParams<- searchSAGAW()
}
}
\author{
Chris Reudenbach
}

