% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagaControl.R
\name{searchSAGAW}
\alias{searchSAGAW}
\title{Searches recursively for existing 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGAW(DL = "C:", quiet = TRUE)
}
\arguments{
\item{DL}{drive letter default is "C:"}

\item{quiet}{boolean  switch for supressing messages default is TRUE}
}
\value{
A dataframe contasining the 'SAGA GIS' root folder(s), the version name(s) and the installation type(s)
}
\description{
Searches recursivley for existing 'SAGA GIS' installation(s) on a given 'Windows' drive
}
\examples{
\dontrun{
#### Examples how to use searchSAGAW 

# get all valid SAGA installation folders and params
searchSAGAW()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
