% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchSAGAW}
\alias{searchSAGAW}
\title{Search for valid 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGAW(DL = "C:", verSelect = FALSE)
}
\arguments{
\item{DL}{drive letter default is "C:"}

\item{verSelect}{boolean default is FALSE. If there is more than one 'SAGA GIS' installation and \code{verSelect} = TRUE the user can select interactively the preferred 'SAGA GIS' version}
}
\value{
A dataframe contasining the 'SAGA GIS' root folder(s), the version name(s) and the installation type(s)
}
\description{
Search for valid 'SAGA GIS' installation(s) on a given 'Windows' drive
}
\examples{
\dontrun{
#### Examples how to use searchSAGAW 

# get all valid SAGA installation folders and params
sagaParams<- searchSAGAW()
}
}
\author{
Chris Reudenbach
}

