% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{searchSAGAX}
\alias{searchSAGAX}
\title{Searches recursively for existing 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGAX(MP = "/usr", ver_select = FALSE, quiet = TRUE)
}
\arguments{
\item{ver_select}{boolean default is FALSE. If there is more than one 'SAGA GIS' installation and \code{ver_select} = TRUE the user can select interactively the preferred 'SAGA GIS' version}

\item{quiet}{boolean  switch for supressing messages default is TRUE}

\item{DL}{drive letter default is "C:"}
}
\value{
A dataframe contasining the 'SAGA GIS' root folder(s), the version name(s) and the installation type(s)
}
\description{
Searches recursivley for existing 'SAGA GIS' installation(s) on a given 'Windows' drive
}
\examples{
\dontrun{
#### Examples how to use searchSAGAX

# get all valid SAGA installation folders and params
sagaParams<- searchSAGAX()
}
}
\author{
Chris Reudenbach
}
