% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagaControl.R
\name{searchSAGAX}
\alias{searchSAGAX}
\title{Searches recursively for existing 'Windows' 'SAGA GIS' installation(s)}
\usage{
searchSAGAX(MP = "/usr", quiet = TRUE)
}
\arguments{
\item{MP}{default mount point is \code{/usr}}

\item{quiet}{boolean  switch for supressing console messages default is TRUE}
}
\value{
A dataframe contasining the 'SAGA GIS' root folder(s), the version name(s) and the installation type(s)
}
\description{
Search for valid 'GRASS GIS' installations at a given 'Linux' mount point
}
\examples{
\dontrun{
#### Examples how to use searchSAGAX

# get all valid SAGA installation folders and params
searchSAGAX()
}
}
\author{
Chris Reudenbach
}
\keyword{internal}
