% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalControls.R
\name{setenvGDAL}
\alias{setenvGDAL}
\title{Usually for internally usage, initializes and set up  access to the 'GDAL' command line interface}
\usage{
setenvGDAL(bin_GDAL = NULL)
}
\arguments{
\item{bin_GDAL}{string contains the path to the 'GDAL' binaries}
}
\value{
Adds 'GDAL' pathes to the environment and creates the variable global string variable \code{gdalCmd}, that contains the path to the 'GDAL' binaries.
}
\description{
Initializes and set up  access to the 'GDAL' command line interface
}
\examples{

run = FALSE
if (run) {
## example for the most common default OSGeo4W64 installation of GDAL
setenvGDAL(bin_GDAL = "C:/OSGeo4W64/bin/",
          root_GDAL = "C:/OSGeo4W64")
}
}
