% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{setOTBEnv}
\alias{setOTBEnv}
\title{setOTBEnv set environ Params of OTB}
\usage{
setOTBEnv(binPathOtb = NULL, rootPathOtb = NULL)
}
\arguments{
\item{binPathOtb}{string contains path to otb binaries}

\item{rootPathOtb}{string contains the full string to call otb launcher}
}
\value{
add otb pathes to the enviroment and creates global variables otbCmd
}
\description{
Provides during the rsession full access to the OTB command line interface
}
\examples{
\dontrun{
## call it for a default OSGeo4W64 oinstallation of SAGA
setOTBEnv()
}
}

