% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initControls.R
\name{setOTBEnv}
\alias{setOTBEnv}
\title{Initializes and set up  access to the 'OTB' command line interface}
\usage{
setOTBEnv(binPathOtb = NULL, rootPathOtb = NULL)
}
\arguments{
\item{binPathOtb}{string contains the path to the 'OTB' binaries}

\item{rootPathOtb}{string contains the full string to the root folder
containing the 'OTB' installation'}
}
\value{
Adds 'OTB' pathes to the enviroment and creates the variable global string variable \code{otbCmd}, that contains the path to the 'OTB' binaries.
}
\description{
Initializes and set up  access to the 'OTB' command line interface
}
\examples{
\dontrun{
## example for the most common default OSGeo4W64 installation of OTB
setOTBEnv(binPathOtb="C:\\\\OSGeo4W64\\\\bin\\\\",rootPathOtb="C:\\\\OSGeo4W64")
}
}

